backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg[binary]
psycopg2-binary
pgvector
python-dotenv
httpx
google-auth
google-auth-oauthlib
google-api-python-client
requests
langchain
openai
aiohttp
python-multipart
pydantic[dotenv]
pydantic-settings
pyjwt
beautifulsoup4
lxml
html5lib
tiktoken
slowapi
redis
bleach
python-jose[cryptography]
passlib[bcrypt]
prometheus-client
sentry-sdk[fastapi]